passing R CMD check again. solved problem related to finalization of Message*

+++ R/wrapper_CodedInputStream.R
@@ -2,7 +2,7 @@
 setGeneric( "ReadRaw", function(object, size ){
 	standardGeneric( "ReadRaw" )
 } )
-setMethod( "ReadRaw", c( object="ZeroCopyInputStream", size = "numeric" ), function(object, size){
+setMethod( "ReadRaw", c( object="ZeroCopyInputStream", size = "integer" ), function(object, size){
 	.Call( "ZeroCopyInputStream_ReadRaw", object@pointer, size, PACKAGE = "RProtoBuf" )
 } )
